+++ tests/unit/example.spec.js
@@ -1,9 +1,9 @@
-import { shallowMount } from "@vue/test-utils";
-import Dumb from "@/components/Dumb.vue";
+import { shallowMount } from '@vue/test-utils'
+import Dumb from '@/components/Dumb.vue'
 
-describe("Dumb.vue", () => {
-  it("is rendered", () => {
-    const wrapper = shallowMount(Dumb);
-    expect(wrapper.is(Dumb)).toBe(true);
-  });
-});
+describe('Dumb.vue', () => {
+  it('is rendered', () => {
+    const wrapper = shallowMount(Dumb)
+    expect(wrapper.is(Dumb)).toBe(true)
+  })
+})